Internal: isolate test utilities in a dedicated crate
Problem: test fixture utils are not available to the client/server
crates.

Solution: extract these functions from the stone-prover crate and add a
new project-wide crate as dev dependency for all crates in the project.
