+++ stone-prover/Cargo.toml
@@ -11,3 +11,5 @@ tempfile = "3.8.1"
 thiserror = "1.0.50"
 tokio = { workspace = true }
 
+[dev-dependencies]
+test-toolkit = { path = "../test-toolkit" }
