+++ stone-prover/src/toolkit.rs
@@ -1,9 +1,5 @@
-#[cfg(test)]
-use std::fs;
 use std::fs::File;
 use std::path::Path;
-#[cfg(test)]
-use std::path::PathBuf;
 
 use serde::de::DeserializeOwned;
 use serde::Serialize;
@@ -25,17 +21,4 @@ pub fn write_json_to_file<T: Serialize, P: AsRef<Path>>(
     let mut file = File::create(path)?;
     serde_json::to_writer(&mut file, &obj)?;
     Ok(())
-}
-
-#[cfg(test)]
-pub fn get_fixture_path(filename: &str) -> PathBuf {
-    Path::new(env!("CARGO_MANIFEST_DIR"))
-        .join("tests/fixtures")
-        .join(filename)
-}
-
-#[cfg(test)]
-pub fn load_fixture(filename: &str) -> String {
-    let fixture_path = get_fixture_path(filename);
-    fs::read_to_string(fixture_path).expect("Failed to read the fixture file")
 }
